Extract src/sidebar.ts — the project sidebar and the mini-rail
First of the evidence-ordered render modules: these are the two surfaces
renderAll() repaints first on every telemetry event, so the Phase-3
profiling pass needs each isolated before it can ask "did this need to
repaint?".

Moved whole (two guarded line ranges): renderSidebar, renderMini,
initProjectDnD, initFileDrop, shellEscapePath, and the two guards that
belong to them — draggingProjects (a mid-drag repaint would destroy the
node the browser is dragging) and reorderGuardUntil.

Two seam-rule-2 hooks, both wanted by the end of a drag: setSidebarSetSort
(manual reorder reasserts manual sort mode) and setSidebarRenderAll.
reorderGuardUntil is read by the global click handler in main.ts, so it
follows state.ts's convention — a live-binding read plus setReorderGuard,
which assigns and nothing else.

IS_MAC / MOD / chord went down to ./dom (seam rule 1): renderMini labels
two buttons with a chord, and so do the palette and the footer, so no one
of them can own it.

Learned, and recorded in PLAN: dom.ts must not touch the DOM at module
scope. Taking the `if (!IS_MAC)` index.html glyph rewrite along with the
constants turned test/tasks.test.ts red with `document is not defined` —
./debug imports ./dom, so vitest pulls it into the node environment via
every logic module that logs. The rewrite is bootstrap and stayed in
main.ts; the constants are fine, node has navigator.userAgent.

No tests: render module, per the Phase-1 rule. Verified mechanical — the
relocated block diffed against `git show HEAD:src/main.ts` differs only by
five `export` keywords and the setReorderGuard seam. main.ts 2,499 -> 2,280.
Gates: tsc --noEmit clean, 352 vitest pass. No Rust touched.
